+++ src/cjs/utils/helpers/VEllipseSectorHelper.d.ts
@@ -6,12 +6,13 @@
  * @modified 2025-04-02 Adding `VEllipseSectorHelper.drawHandleLines`.
  * @modified 2025-04-07 Modifying the calculation of `startAngle` and `endAngle` from the rotation control point: wrapping result into [0,TWO_PI).
  * @modified 2025-04-09 Adding the `VEllipseSectorHelper.destroy` method.
+ * @modified 2025-05-05 `VEllipseSectorHelper` is now implementing `IShapeInteractionHelper`.
  * @version  1.1.0
  */
 import { VEllipseSector } from "../../VEllipseSector";
 import { Vertex } from "../../Vertex";
-import { DrawLib } from "../../interfaces";
-export declare class VEllipseSectorHelper {
+import { DrawLib, IShapeInteractionHelper } from "../../interfaces";
+export declare class VEllipseSectorHelper implements IShapeInteractionHelper {
     private sector;
     private startAngleControlPoint;
     private endAngleControlPoint;