+++ src/cjs/utils/helpers/VEllipseSectorHelper.js
@@ -7,6 +7,7 @@
  * @modified 2025-04-02 Adding `VEllipseSectorHelper.drawHandleLines`.
  * @modified 2025-04-07 Modifying the calculation of `startAngle` and `endAngle` from the rotation control point: wrapping result into [0,TWO_PI).
  * @modified 2025-04-09 Adding the `VEllipseSectorHelper.destroy` method.
+ * @modified 2025-05-05 `VEllipseSectorHelper` is now implementing `IShapeInteractionHelper`.
  * @version  1.1.0
  */
 Object.defineProperty(exports, "__esModule", { value: true });