+++ src/ts/utils/helpers/BezierPathInteractionHelper.ts
@@ -25,7 +25,9 @@
  * @modified 2025-04-14 Added the `BezierPathInteractionHelper.drawHandleLines` method.
  * @modified 2025-04-14 Fixing correct event types for touch events in `BezierPathInteractionHelper`.
  * @modified 2025-04-14 BezierPathInteractionHelper: Changed default value of `HelperOptions.autoAdjustPaths` from `true` to `false`.
- * @version  1.2.0
+ * @modified 2025-05-05 Added optional params `draw` and `fill` to BezierPathInteractionHelper.drawHandleLines` method.
+ * @modified 2025-05-05 Class `BezierPathInteractionHelper` now implementing `IShapeInteractionHelper`.
+ * @version  1.2.1
  *
  * @file BezierPathInteractionHelper
  * @public
@@ -42,7 +44,7 @@ import { PlotBoilerplate } from "../../PlotBoilerplate";
 import { VertListener } from "../../VertexListeners";
 import { Vertex } from "../../Vertex";
 import { XMouseEvent } from "../../MouseHandler";
-import { DrawLib, XYCoords } from "../../interfaces";
+import { DrawLib, IShapeInteractionHelper, XYCoords } from "../../interfaces";
 import { AFTouchEvent } from "alloyfinger-typescript/src/cjs/alloy_finger";
 
 /**
@@ -91,7 +93,7 @@ interface HelperOptions {
  *
  * @public
  **/
-export class BezierPathInteractionHelper {
+export class BezierPathInteractionHelper implements IShapeInteractionHelper {
   /**
    * @member {PlotBoilerplate} pb
    * @memberof BezierPathInteractionHelper
@@ -273,8 +275,12 @@ export class BezierPathInteractionHelper {
   /**
    * Draw grey handle lines.
    *
+   * @param {DrawLib<any>} draw - (optional) The draw library to use. If not provided then `pb.draw` will be used.
+   * @param {DrawLib<any>} fill - (optional) The fill library to use. If not provided then `pb.fill` will be used.
    */
-  drawHandleLines() {
+  drawHandleLines(draw?: DrawLib<any>, fill?: DrawLib<any>) {
+    draw = draw || this.pb.draw;
+    fill = fill || this.pb.fill;
     this.paths.forEach((path: BezierPath) => {
       path.bezierCurves.forEach((curve: CubicBezierCurve) => {
         this.pb.draw.line(curve.startPoint, curve.startControlPoint, "rgba(64,192,128,0.333)", 1.0, {